refactor(audiobooks): split AudiobookService into focused sub-files (ARCH-5)
Split the 2691-line service.go god file into 6 thematically-grouped
files, all within the same `audiobooks` package (no API changes):

- service.go (171 lines): interfaces, struct, constructor, Set* wires,
  InvalidateBookCaches, deref/cmp utility functions
- service_types.go (158 lines): all request/response/filter types,
  PerUserFieldNames, strippedMemdbFields, splitFieldFilters helpers
- service_filtering.go (831 lines): sort map, filter predicates, pushdown
  builders, bookSummary converters, file aggregation helpers
- service_query.go (510 lines): GetAudiobooks, CountAudiobooksFiltered,
  FetchBookFilesForBooks, Enrich*, searchWithBleve
- service_single.go (470 lines): GetAudiobook, GetAudiobookTags,
  extractBookFileMetadata, GetDuplicate/SoftDeleted/Purge/Restore/Count
- service_mutation.go (554 lines): UpdateAudiobook, DeleteAudiobook,
  collectITunesPIDsForBook, enqueueITunesRemovesForBook, ApplyOverrideToPayload
- service_tags.go (93 lines): all ListAllUserTags / GetBookUserTags /
  SetBookUserTags / AddBookUserTag / RemoveBookUserTag / BatchUpdateUserTags

Build: GOEXPERIMENT=jsonv2 go build ./... passes clean.
Tests: GOEXPERIMENT=jsonv2 go test ./internal/audiobooks/... passes (60s).
